refactor: move EnvItem class to appropriate reference location
N
