+++ apiserver/paasng/paasng/infras/legacydb/adaptors.py
@@ -29,13 +29,14 @@ from sqlalchemy import func, or_
 from sqlalchemy.exc import IntegrityError as SqlIntegrityError
 from sqlalchemy.orm import Session
 
-from paasng.platform.engine.constants import ConfigVarEnvName
-from paasng.platform.applications.exceptions import IntegrityError
-from paasng.infras.legacydb import models as legacy_models
-from paasng.platform.mgrlegacy.constants import LegacyAppState
-from paasng.infras.oauth2.models import OAuth2Client
-from paasng.accessories.publish.sync_market.constant import EnvItem, SaaSPackageInfo
+from paasng.accessories.publish.sync_market.constant import SaaSPackageInfo
 from paasng.accessories.publish.sync_market.models import TagData
+from paasng.infras.legacydb import models as legacy_models
+from paasng.infras.legacydb.entities import EnvItem
+from paasng.infras.oauth2.models import OAuth2Client
+from paasng.platform.applications.exceptions import IntegrityError
+from paasng.platform.engine.constants import ConfigVarEnvName
+from paasng.platform.mgrlegacy.constants import LegacyAppState
 
 logger = logging.getLogger(__name__)
